调整结构

+++ docker-install-debian.sh
@@ -48,7 +48,9 @@ sudo systemctl restart docker
 #安装docker compose
 
 #卸载旧版本
-sudo rm /usr/local/bin/docker-compose
+if [ -f /usr/local/bin/docker-compose ]; then
+  sudo rm /usr/local/bin/docker-compose
+fi
 
 #安装新版本
 curl -SL https://github.com/docker/compose/releases/download/v2.4.1/docker-compose-linux-x86_64 -o /usr/local/bin/docker-compose
